requirements.txt
amqp==5.0.5
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.6
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.4.0
kombu==5.0.2
model-bakery==1.2.1
#mysqlclient==2.0.1
Pillow==8.1.1
prompt-toolkit==3.0.16
pytz==2021.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
gunicorn==20.0.4
psycopg2==2.8.6
whitenoise==5.2.0
